requirements.txt
aniso8601==1.2.1
apsw==3.8.11.1.post1
BeautifulSoup==3.2.1
beautifulsoup4==4.5.1
chardet==2.3.0
CherryPy==3.5.0
click==6.7
cryptography==1.5
cssselect==0.9.2
cssutils==1.0
defusedxml==0.4.1
dnspython==1.14.0
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
feedparser==5.1.3
file-magic==0.3.0
Flask==0.12.1
Flask-HTTPAuth==3.2.3
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
Flask-SocketIO==2.8.6
gevent==1.2.1
greenlet==0.4.12
html5lib==0.999
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
lxml==3.6.4
Markdown==2.6.7
MarkupSafe==1.0
marshmallow==3.0.0b2
mechanize==0.2.5
mysql-connector-python==2.1.3
mysql-utilities==1.6.3
mysqlclient==1.3.7
netifaces==0.10.4
numpy==1.11.1rc1
paramiko==2.0.0
pexpect==4.2.0
Pillow==3.3.1
Pivy==0.5.0
ptyprocess==0.5.1
py==1.4.34
pyasn1==0.1.9
pycrypto==2.6.1
pycurl==7.43.0
pygdbmi==0.7.3.3
Pygments==2.2.0
pygobject==3.22.0
PyJWT==1.5.2
pyliblzma==0.5.3
pymongo==3.4.0
PyMySQL==0.7.11
pyodbc==3.0.10
pyparsing==2.1.8
pypugjs==4.2.0
pysqlite==2.7.0
pytest==3.1.3
python-apt==1.1.0b5
python-dateutil==2.4.2
python-engineio==1.3.1
python-socketio==1.7.4
pytz==2017.2
PyYAML==3.11
repoze.lru==0.6
Routes==2.2
scipy==0.19.0
six==1.10.0
SOAPpy==0.12.22
Theano==0.9.0
uTidylib==0.3
virtualenv==15.1.0
WebOb==1.6.1
Werkzeug==0.12.1
wstools==0.4.3
